# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  displayName: string;
  avatarColor: string;
  createdAt: string;
}

export interface SignupRequest {
  /**
   * @minLength 3
   * @maxLength 20
   */
  username: string;
  /** @minLength 6 */
  password: string;
  /** @maxLength 30 */
  displayName: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export interface UpdateUserRequest {
  /** @maxLength 30 */
  displayName?: string;
  avatarColor?: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}